Create add function
